[IL2Asm] Merge CALLVIRT and CALL together, reducing code duplication.
Include a MethodSpecLayout with each MethodHeader so that we can find any generics if applicable.
Add basic support for generics to the 32b assembler.
Add basic support for the HASTHIS flag by offsetting arguments and adding the correct number of bytes to the return type.
Add 'PtrToObject' plug (which is just a nop) in prep for having a heap.
Do a better job of including the generic type name in the method name.