[OS] Place a valid VBR (in the form of a jump instruction) into the first sector of the FAT32 filesystem.
Take advantage of this and make stage 1 bootloader completely agnostic of the partition type.  It always just loads the first 8 sectors into memory and jumps to 0x9000.  This should allow alternate file systems to work, with only adjustments to stage 2.
Clean up some of the code and variable names in stage 1